+++ project_config/webpack/loaders/html_loader.js
@@ -1,6 +1,4 @@
 
-const cloneDeep = require('lodash/cloneDeep');
-
 /**
  * Make `html-loader`.
  *
@@ -8,10 +6,12 @@ const cloneDeep = require('lodash/cloneDeep');
  * @returns {{loader: string, test: RegExp, options: {attrs: string[], interpolate: string}}}
  */
 module.exports = config => {
-    const loader = config.htmlLoader ? cloneDeep(config.htmlLoader) : {};
-
-    loader.loader = 'html-loader';
-    loader.test = /\.html$/;
-
-    return loader;
+    return {
+        loader: 'html-loader',
+        test: /\.html$/,
+        options: {
+            attrs: ['img:src', 'link:href'],
+            interpolate: 'require'
+        }
+    };
 };