+++ project_config/webpack/loaders/url_loader.js
@@ -9,16 +9,17 @@ const cloneDeep = require('lodash/cloneDeep');
  * @returns {{loader: string, options: {limit: number}, test: RegExp}}
  */
 module.exports = (config, isBuild) => {
-    const loader = config.urlLoader ? cloneDeep(config.urlLoader) : {};
-
-    !loader.options && (loader.options = {});
-
-    loader.loader = 'url-loader';
-    loader.test = new RegExp(`\.(${config.fileLoaderSuffixes.join('|')})$`);
+    let options = {
+        limit: 8192
+    };
 
     if (isBuild) {
-        loader.options.publicPath = config.staticServerDir + config.basePaths.webPrefix + '/dist/'
+        options.publicPath = config.staticServerDir + config.basePaths.webPrefix + '/dist/'
     }
 
-    return loader;
+    return {
+        loader: 'url-loader',
+        options: options,
+        test: new RegExp(`\.(${config.fileLoaderSuffixes.join('|')})$`)
+    }
 };